# examples/reference-app/internal/smilesim/api/smilesim-server.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

// Defines values for SmilesimJobStatusStatus.
const (
	SmilesimJobStatusStatusCancelled  SmilesimJobStatusStatus = "cancelled"
	SmilesimJobStatusStatusDeadLetter SmilesimJobStatusStatus = "dead_letter"
	SmilesimJobStatusStatusPending    SmilesimJobStatusStatus = "pending"
	SmilesimJobStatusStatusRetrying   SmilesimJobStatusStatus = "retrying"
	SmilesimJobStatusStatusRunning    SmilesimJobStatusStatus = "running"
	SmilesimJobStatusStatusSucceeded  SmilesimJobStatusStatus = "succeeded"
)

// Valid indicates whether the value is a known member of the SmilesimJobStatusStatus enum.
func (e SmilesimJobStatusStatus) Valid() bool {
	switch e {
	case SmilesimJobStatusStatusCancelled:
		return true
	case SmilesimJobStatusStatusDeadLetter:
		return true
	case SmilesimJobStatusStatusPending:
		return true
	case SmilesimJobStatusStatusRetrying:
		return true
	case SmilesimJobStatusStatusRunning:
		return true
	case SmilesimJobStatusStatusSucceeded:
		return true
	default:
		return false
	}
}

// Defines values for SmilesimSimulationStatus.
const (
	SmilesimSimulationStatusCancelled  SmilesimSimulationStatus = "cancelled"
	SmilesimSimulationStatusDeadLetter SmilesimSimulationStatus = "dead_letter"
	SmilesimSimulationStatusPending    SmilesimSimulationStatus = "pending"
	SmilesimSimulationStatusRetrying   SmilesimSimulationStatus = "retrying"
	SmilesimSimulationStatusRunning    SmilesimSimulationStatus = "running"
	SmilesimSimulationStatusSucceeded  SmilesimSimulationStatus = "succeeded"
)

// Valid indicates whether the value is a known member of the SmilesimSimulationStatus enum.
func (e SmilesimSimulationStatus) Valid() bool {
	switch e {
	case SmilesimSimulationStatusCancelled:
		return true
	case SmilesimSimulationStatusDeadLetter:
		return true
	case SmilesimSimulationStatusPending:
		return true
	case SmilesimSimulationStatusRetrying:
		return true
	case SmilesimSimulationStatusRunning:
		return true
	case SmilesimSimulationStatusSucceeded:
		return true
	default:
		return false
	}
}

// Defines values for SmilesimSimulationOptionsSmileStyle.
const (
	SmilesimSimulationOptionsSmileStyleBright  SmilesimSimulationOptionsSmileStyle = "bright"
	SmilesimSimulationOptionsSmileStyleNatural SmilesimSimulationOptionsSmileStyle = "natural"
	SmilesimSimulationOptionsSmileStyleSubtle  SmilesimSimulationOptionsSmileStyle = "subtle"
)

// Valid indicates whether the value is a known member of the SmilesimSimulationOptionsSmileStyle enum.
func (e SmilesimSimulationOptionsSmileStyle) Valid() bool {
	switch e {
	case SmilesimSimulationOptionsSmileStyleBright:
		return true
	case SmilesimSimulationOptionsSmileStyleNatural:
		return true
	case SmilesimSimulationOptionsSmileStyleSubtle:
		return true
	default:
		return false
	}
}

// Defines values for SmilesimSimulationOptionsToothShade.
const (
	SmilesimSimulationOptionsToothShadeNatural    SmilesimSimulationOptionsToothShade = "natural"
	SmilesimSimulationOptionsToothShadeUltraWhite SmilesimSimulationOptionsToothShade = "ultra-white"
	SmilesimSimulationOptionsToothShadeWhite      SmilesimSimulationOptionsToothShade = "white"
)

// Valid indicates whether the value is a known member of the SmilesimSimulationOptionsToothShade enum.
func (e SmilesimSimulationOptionsToothShade) Valid() bool {
	switch e {
	case SmilesimSimulationOptionsToothShadeNatural:
		return true
	case SmilesimSimulationOptionsToothShadeUltraWhite:
		return true
	case SmilesimSimulationOptionsToothShadeWhite:
		return true
	default:
		return false
	}
}

// SmilesimError The structured {code, params} error envelope every speed API returns instead of localized text: code is a stable machine-readable identifier naming exactly one failure, and params carries the structured per-code details when the failure has any. The server never renders a message -- a client maps each code to its own locale's text.
type SmilesimError struct {
	// Code Example: smilesim.invalid_request_body
	Code   *string                 `json:"code,omitempty"`
	Params *map[string]interface{} `json:"params,omitempty"`
}

// SmilesimJobRef The simulate route's 202 answer: a reference to the async job, which has not run yet.
type SmilesimJobRef struct {
	// JobID The enqueued go/jobs job's id -- the value the job-status route's {jobID} path parameter takes.
	JobID string `json:"job_id"`
}

// SmilesimJobStatus The job-status route's 200 answer for one job.
type SmilesimJobStatus struct {
	// Error The job's recorded failure message, present when the job carries one (dead-lettered, cancelled after a failed attempt, or retrying).
	Error *string `json:"error,omitempty"`

	// Options One simulation's parameterized option set, shared by the simulate route's request body and by every response that echoes the effective options a simulation was generated with.
	Options *SmilesimSimulationOptions `json:"options,omitempty"`

	// OutputObjectID The generated image's go/storage object id, present once the job has succeeded.
	OutputObjectID *string `json:"output_object_id,omitempty"`

	// Status The job's live go/jobs status.
	Status SmilesimJobStatusStatus `json:"status"`

	// Usage The real vendor usage a succeeded job recorded.
	Usage *SmilesimUsage `json:"usage,omitempty"`
}

// SmilesimJobStatusStatus The job's live go/jobs status.
type SmilesimJobStatusStatus string

// SmilesimPhotoSimulations The per-photo enumeration route's 200 answer.
type SmilesimPhotoSimulations struct {
	// Simulations The photo's simulations, newest first.
	Simulations []SmilesimSimulation `json:"simulations"`
}

// SmilesimSimulateRequest The simulate route's request body. Every options field is a pointer-shaped optional: an omitted field inherits the service's documented default, while an explicitly present field -- including an out-of-range strength of 0 -- is passed through verbatim and refused by the service's own validation rather than silently replaced by a default.
type SmilesimSimulateRequest struct {
	// Options One simulation's parameterized option set, shared by the simulate route's request body and by every response that echoes the effective options a simulation was generated with.
	Options *SmilesimSimulationOptions `json:"options,omitempty"`

	// PhotoObjectID The completed go/storage object id of the patient photo to simulate.
	PhotoObjectID string `json:"photo_object_id"`

	// RecipientUserID The user id of the recipient to notify once the job reaches a terminal status. Omitted (or empty) enqueues the job with no notification.
	RecipientUserID *string `json:"recipient_user_id,omitempty"`
}

// SmilesimSimulation One simulation in the per-photo enumeration.
type SmilesimSimulation struct {
	// CreatedAt When the generation was requested, whole seconds, UTC.
	CreatedAt time.Time `json:"created_at"`

	// Error The job's recorded failure message, present only when the job carries one.
	Error *string `json:"error,omitempty"`

	// JobID The generation job's id.
	JobID string `json:"job_id"`

	// Options One simulation's parameterized option set, shared by the simulate route's request body and by every response that echoes the effective options a simulation was generated with.
	Options SmilesimSimulationOptions `json:"options"`

	// OutputObjectID The generated image's go/storage object id, present only when the job has succeeded and its result names one.
	OutputObjectID *string `json:"output_object_id,omitempty"`

	// PhotoObjectID The go/storage object id of the photo the simulation was generated from -- the enumeration's grouping key, echoed per entry for a caller rendering a row on its own.
	PhotoObjectID string `json:"photo_object_id"`

	// Status The job's live status, read from the queue at call time -- never a snapshot stored by the service, so an enumeration always reflects what a job-status poll would answer right now.
	Status SmilesimSimulationStatus `json:"status"`
}

// SmilesimSimulationStatus The job's live status, read from the queue at call time -- never a snapshot stored by the service, so an enumeration always reflects what a job-status poll would answer right now.
type SmilesimSimulationStatus string

// SmilesimSimulationContent The simulation-content read's 200 answer for one generated image.
type SmilesimSimulationContent struct {
	// ContentBase64 The stored bytes, base64-encoded -- the same JSON-carried transport the cases surface's photo-content route uses.
	ContentBase64 string `json:"content_base64"`

	// MediaType The media type go/storage's probe assigned to the stored bytes at complete time.
	MediaType string `json:"media_type"`
}

// SmilesimSimulationOptions One simulation's parameterized option set, shared by the simulate route's request body and by every response that echoes the effective options a simulation was generated with.
type SmilesimSimulationOptions struct {
	// SmileStyle The character of the simulated smile. Defaults to natural when omitted.
	SmileStyle *SmilesimSimulationOptionsSmileStyle `json:"smile_style,omitempty"`

	// Strength How strongly the transformation moves the photo toward the requested smile: the lower bound is exclusive (0 is a billed no-op and is refused) and 1 is the complete simulated smile. Defaults to full strength when omitted.
	Strength *float64 `json:"strength,omitempty"`

	// ToothShade The tooth color the simulated smile is asked to show. Defaults to natural when omitted.
	ToothShade *SmilesimSimulationOptionsToothShade `json:"tooth_shade,omitempty"`
}

// SmilesimSimulationOptionsSmileStyle The character of the simulated smile. Defaults to natural when omitted.
type SmilesimSimulationOptionsSmileStyle string

// SmilesimSimulationOptionsToothShade The tooth color the simulated smile is asked to show. Defaults to natural when omitted.
type SmilesimSimulationOptionsToothShade string

// SmilesimUsage The real vendor usage a succeeded job recorded.
type SmilesimUsage struct {
	// ImageCount How many images the vendor generated for the job.
	ImageCount int `json:"image_count"`

	// ResolutionTier The resolution tier the vendor generated at (its normalized size string).
	ResolutionTier string `json:"resolution_tier"`

	// Steps The inference step count the vendor reported.
	Steps int `json:"steps"`
}

// SmilesimSimulateJSONRequestBody defines body for SmilesimSimulate for application/json ContentType.
type SmilesimSimulateJSONRequestBody = SmilesimSimulateRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// SmilesimGetJob Poll one smile-simulation job's live status.
	// (GET /api/v1/smile-simulation/jobs/{jobID})
	SmilesimGetJob(w http.ResponseWriter, r *http.Request, jobID string)
	// SmilesimListPhotoSimulations List every simulation generated from one photo.
	// (GET /api/v1/smile-simulation/photos/{photoObjectID}/simulations)
	SmilesimListPhotoSimulations(w http.ResponseWriter, r *http.Request, photoObjectID string)
	// SmilesimGetSimulationContent Read one simulation's generated image content.
	// (GET /api/v1/smile-simulation/photos/{photoObjectID}/simulations/{jobID}/content)
	SmilesimGetSimulationContent(w http.ResponseWriter, r *http.Request, photoObjectID string, jobID string)
	// SmilesimSimulate Enqueue one smile-simulation job for a completed photo.
	// (POST /api/v1/smile-simulation/simulate)
	SmilesimSimulate(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// SmilesimGetJob operation middleware
func (siw *ServerInterfaceWrapper) SmilesimGetJob(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "jobID" -------------
	var jobID string

	err = runtime.BindStyledParameterWithOptions("simple", "jobID", r.PathValue("jobID"), &jobID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "jobID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SmilesimGetJob(w, r, jobID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SmilesimListPhotoSimulations operation middleware
func (siw *ServerInterfaceWrapper) SmilesimListPhotoSimulations(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "photoObjectID" -------------
	var photoObjectID string

	err = runtime.BindStyledParameterWithOptions("simple", "photoObjectID", r.PathValue("photoObjectID"), &photoObjectID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "photoObjectID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SmilesimListPhotoSimulations(w, r, photoObjectID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SmilesimGetSimulationContent operation middleware
func (siw *ServerInterfaceWrapper) SmilesimGetSimulationContent(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "photoObjectID" -------------
	var photoObjectID string

	err = runtime.BindStyledParameterWithOptions("simple", "photoObjectID", r.PathValue("photoObjectID"), &photoObjectID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "photoObjectID", Err: err})
		return
	}

	// ------------- Path parameter "jobID" -------------
	var jobID string

	err = runtime.BindStyledParameterWithOptions("simple", "jobID", r.PathValue("jobID"), &jobID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "jobID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SmilesimGetSimulationContent(w, r, photoObjectID, jobID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SmilesimSimulate operation middleware
func (siw *ServerInterfaceWrapper) SmilesimSimulate(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SmilesimSimulate(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/v1/smile-simulation/simulate", wrapper.SmilesimSimulate)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/v1/smile-simulation/jobs/{jobID}", wrapper.SmilesimGetJob)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/v1/smile-simulation/photos/{photoObjectID}/simulations", wrapper.SmilesimListPhotoSimulations)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/v1/smile-simulation/photos/{photoObjectID}/simulations/{jobID}/content", wrapper.SmilesimGetSimulationContent)

	return m
}