# examples/reference-app/web/app-openapi.yaml
openapi: 3.0.3
info:
  title: reference-app
  description: "The reference-app application OpenAPI document: the redocly join of the application own three fragments (notes, cases and smilesim), regenerated by `task api:gen:app` and committed. Each fragment stays authoritative for its own operations; this merged document is the app-owned SDK generation input and the linted whole-surface view."
  version: 0.0.0
...
paths:
...
  /api/v1/smile-simulation/simulate:
    servers: []
    post:
      operationId: smilesim_simulate
      tags:
        - smilesim
      summary: Enqueue one smile-simulation job for a completed photo.
      description: >-
        The photo must be an existing, completed go/storage object of the
        caller's own tenant; the job itself is async -- the 202 answer carries
        only its job id, and the job-status route below is how its outcome is
        observed. Options are optional: an absent options object (or an absent
        field inside it) inherits the service's documented default for that
        dimension, while an explicitly present out-of-vocabulary or out-of-range
        value is refused with its coded smilesim.* 400 before anything is
        reserved or enqueued. A recipient_user_id, when given, receives a
        completion notification once the job reaches a terminal status. There is
        deliberately no tenant_id field anywhere on this request: the tenant is
        the one tenancy.Middleware already resolved into the request context --
        never a caller-supplied value.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmilesimSimulateRequest'
      responses:
        '202':
          description: >-
            The simulation job was enqueued; poll the job-status route for its
            outcome.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimJobRef'
        '400':
          description: >-
            The request body was missing or malformed, photo_object_id was
            absent, or an explicitly present option was out of vocabulary or out
            of range (smilesim.unsupported_smile_style,
            smilesim.unsupported_tooth_shade, smilesim.strength_out_of_range).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
        '401':
          description: The access token was missing or invalid (authn.token_invalid).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
        '403':
          description: >-
            The request carried no tenant -- an anonymous request to a
            tenant-scoped route (tenancy.tenant_unresolved).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
        '409':
          description: >-
            The credit reservation was refused: the tenant's balance cannot
            cover one simulation's cost (billing's coded 409).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
        '500':
          description: An unexpected server failure (smilesim.internal_error).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
  /api/v1/smile-simulation/jobs/{jobID}:
    servers: []
    get:
      operationId: smilesim_getJob
      tags:
        - smilesim
      summary: Poll one smile-simulation job's live status.
      description: >-
        Answers the same jobs.Queue this app shares with go/ai-gateway: the
        status is the job's live status, an unknown or another tenant's job id
        answers 404 exactly like an unknown one. Once the job has succeeded, the
        answer carries the generated image's go/storage object id and the real
        vendor usage the job recorded; when a durable per-photo record exists
        for the job, it also echoes the effective options the simulation was
        generated with.
      parameters:
        - name: jobID
          in: path
          required: true
          schema:
            type: string
          description: The job id the simulate route's 202 answer carried.
      responses:
        '200':
          description: The job's live status, plus outcome fields once it is terminal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimJobStatus'
        '401':
          description: The access token was missing or invalid (authn.token_invalid).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
        '403':
          description: >-
            The request carried no tenant -- an anonymous request to a
            tenant-scoped route (tenancy.tenant_unresolved).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
        '404':
          description: >-
            No such job under the caller's tenant, or the job belongs to another
            tenant (jobs.job_not_found -- the same answer either way).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
        '500':
          description: An unexpected server failure (smilesim.internal_error).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
  /api/v1/smile-simulation/photos/{photoObjectID}/simulations:
    servers: []
    get:
      operationId: smilesim_listPhotoSimulations
      tags:
        - smilesim
      summary: List every simulation generated from one photo.
      description: >-
        Lists every simulation generated from the photo under the caller's
        tenant, newest first -- the P3 gallery's data source for one case photo.
        Each entry carries the effective options it was generated with, its live
        status and, once the job succeeded, its output object id. An unknown
        photo id is an empty list, not an error: this route reads the durable
        per-photo result index, which simply holds no rows for a photo nothing
        was generated from.
      parameters:
        - name: photoObjectID
          in: path
          required: true
          schema:
            type: string
          description: >-
            The go/storage object id of the photo the simulations were generated
            from.
      responses:
        '200':
          description: The photo's simulations, newest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimPhotoSimulations'
        '401':
          description: The access token was missing or invalid (authn.token_invalid).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
        '403':
          description: >-
            The request carried no tenant -- an anonymous request to a
            tenant-scoped route (tenancy.tenant_unresolved).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
        '500':
          description: An unexpected server failure (smilesim.internal_error).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
  /api/v1/smile-simulation/photos/{photoObjectID}/simulations/{jobID}/content:
    servers: []
    get:
      operationId: smilesim_getSimulationContent
      tags:
        - smilesim
      summary: Read one simulation's generated image content.
      description: >-
        Serves the stored bytes of the image one simulation job generated,
        base64-encoded in JSON with the media type go/storage's probe assigned
        -- the same JSON-only transport the cases surface's photo-content route
        uses, so the comparison view can render the result beside the original
        photo. The simulation is addressed the way the per-photo enumeration
        lists it: by its generation job's id under the photo it was generated
        from. The job-status route stays the poll for the outcome; this route is
        the read of the bytes once the job has succeeded. A job the photo has no
        record of under the caller's tenant answers 404
        (smilesim.simulation_not_found), a job that has not succeeded answers
        404 (smilesim.output_not_ready), and a succeeded simulation whose stored
        bytes no longer exist answers 404 (smilesim.output_not_found) -- each
        refusal the same outward shape, never probing whether another tenant's
        object still exists.
      parameters:
        - name: photoObjectID
          in: path
          required: true
          schema:
            type: string
          description: >-
            The go/storage object id of the photo the simulation was generated
            from.
        - name: jobID
          in: path
          required: true
          schema:
            type: string
          description: The generation job's id, as the per-photo enumeration lists it.
      responses:
        '200':
          description: >-
            The generated image's stored bytes, base64-encoded with the
            probe-assigned media type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimSimulationContent'
        '401':
          description: The access token was missing or invalid (authn.token_invalid).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
        '403':
          description: >-
            The request carried no tenant -- an anonymous request to a
            tenant-scoped route (tenancy.tenant_unresolved).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
        '404':
          description: >-
            The photo has no such simulation under the caller's tenant
            (smilesim.simulation_not_found), the job has not succeeded yet
            (smilesim.output_not_ready), or the stored bytes of a succeeded
            simulation no longer exist (smilesim.output_not_found).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
        '500':
          description: An unexpected server failure (smilesim.internal_error).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmilesimError'
components:
  schemas:
...
    SmilesimSimulateRequest:
      type: object
      required:
        - photo_object_id
      properties:
        photo_object_id:
          type: string
          minLength: 1
          description: The completed go/storage object id of the patient photo to simulate.
        recipient_user_id:
          type: string
          description: >-
            The user id of the recipient to notify once the job reaches a
            terminal status. Omitted (or empty) enqueues the job with no
            notification.
        options:
          $ref: '#/components/schemas/SmilesimSimulationOptions'
      description: >-
        The simulate route's request body. Every options field is a
        pointer-shaped optional: an omitted field inherits the service's
        documented default, while an explicitly present field -- including an
        out-of-range strength of 0 -- is passed through verbatim and refused by
        the service's own validation rather than silently replaced by a default.
    SmilesimSimulationOptions:
      type: object
      properties:
        smile_style:
          type: string
          enum:
            - subtle
            - natural
            - bright
          description: >-
            The character of the simulated smile. Defaults to natural when
            omitted.
        tooth_shade:
          type: string
          enum:
            - natural
            - white
            - ultra-white
          description: >-
            The tooth color the simulated smile is asked to show. Defaults to
            natural when omitted.
        strength:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
          maximum: 1
          description: >-
            How strongly the transformation moves the photo toward the requested
            smile: the lower bound is exclusive (0 is a billed no-op and is
            refused) and 1 is the complete simulated smile. Defaults to full
            strength when omitted.
      description: >-
        One simulation's parameterized option set, shared by the simulate
        route's request body and by every response that echoes the effective
        options a simulation was generated with.
    SmilesimJobRef:
      type: object
      required:
        - job_id
      properties:
        job_id:
          type: string
          description: >-
            The enqueued go/jobs job's id -- the value the job-status route's
            {jobID} path parameter takes.
      description: >-
        The simulate route's 202 answer: a reference to the async job, which has
        not run yet.
...
    SmilesimJobStatus:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - pending
            - running
            - retrying
            - succeeded
            - dead_letter
            - cancelled
          description: The job's live go/jobs status.
        options:
          $ref: '#/components/schemas/SmilesimSimulationOptions'
          description: >-
            The effective options the job was generated with, present when a
            durable per-photo record exists for it.
        output_object_id:
          type: string
          description: >-
            The generated image's go/storage object id, present once the job has
            succeeded.
        usage:
          $ref: '#/components/schemas/SmilesimUsage'
          description: >-
            The real vendor usage the job recorded, present once it has
            succeeded.
        error:
          type: string
          description: >-
            The job's recorded failure message, present when the job carries one
            (dead-lettered, cancelled after a failed attempt, or retrying).
      description: The job-status route's 200 answer for one job.
...
    SmilesimPhotoSimulations:
      type: object
      required:
        - simulations
      properties:
        simulations:
          type: array
          items:
            $ref: '#/components/schemas/SmilesimSimulation'
          description: The photo's simulations, newest first.
      description: The per-photo enumeration route's 200 answer.
    SmilesimSimulationContent:
      type: object
      required:
        - media_type
        - content_base64
      properties:
        media_type:
          type: string
          description: >-
            The media type go/storage's probe assigned to the stored bytes at
            complete time.
        content_base64:
          type: string
          description: >-
            The stored bytes, base64-encoded -- the same JSON-carried transport
            the cases surface's photo-content route uses.
      description: The simulation-content read's 200 answer for one generated image.
    SmilesimError:
      type: object
      description: >-
        The structured {code, params} error envelope every speed API returns
        instead of localized text: code is a stable machine-readable identifier
        naming exactly one failure, and params carries the structured per-code
        details when the failure has any. The server never renders a message --
        a client maps each code to its own locale's text.
      properties:
        code:
          type: string
          example: smilesim.invalid_request_body
        params:
          type: object
          additionalProperties: true
...